refactor: modularize physical attack mechanics
