+++ src/utils/gameplay/physicalAttacksTypes.ts
@@ -1,0 +1,94 @@
+import { IComponentDamageState, IUnitGameState } from '@/types/gameplay';
+import { CombatLocation } from '@/types/gameplay';
+
+export type PhysicalAttackType =
+  | 'punch'
+  | 'kick'
+  | 'charge'
+  | 'dfa'
+  | 'push'
+  | 'hatchet'
+  | 'sword'
+  | 'mace';
+
+export interface IPhysicalAttackInput {
+  readonly attackerTonnage: number;
+  readonly pilotingSkill: number;
+  readonly componentDamage: IComponentDamageState;
+  readonly attackType: PhysicalAttackType;
+  readonly arm?: 'left' | 'right';
+  readonly hexesMoved?: number;
+  readonly heat?: number;
+  readonly hasTSM?: boolean;
+  readonly isUnderwater?: boolean;
+  readonly weaponsFiredFromArm?: readonly string[];
+  readonly attackerProne?: boolean;
+  readonly targetTonnage?: number;
+}
+
+export interface IPhysicalToHitResult {
+  readonly baseToHit: number;
+  readonly finalToHit: number;
+  readonly modifiers: readonly IPhysicalModifier[];
+  readonly allowed: boolean;
+  readonly restrictionReason?: string;
+}
+
+export interface IPhysicalModifier {
+  readonly name: string;
+  readonly value: number;
+  readonly source: string;
+}
+
+export interface IPhysicalDamageResult {
+  readonly targetDamage: number;
+  readonly attackerDamage: number;
+  readonly attackerLegDamagePerLeg: number;
+  readonly targetPSR: boolean;
+  readonly attackerPSR: boolean;
+  readonly attackerPSRModifier: number;
+  readonly hitTable: 'punch' | 'kick';
+  readonly targetDisplaced: boolean;
+}
+
+export interface IPhysicalAttackResult {
+  readonly attackType: PhysicalAttackType;
+  readonly toHitNumber: number;
+  readonly roll: number;
+  readonly hit: boolean;
+  readonly targetDamage: number;
+  readonly attackerDamage: number;
+  readonly attackerLegDamagePerLeg: number;
+  readonly targetPSR: boolean;
+  readonly attackerPSR: boolean;
+  readonly attackerPSRModifier: number;
+  readonly hitLocation?: CombatLocation;
+  readonly targetDisplaced: boolean;
+}
+
+export interface IPhysicalAttackRestriction {
+  allowed: boolean;
+  reason?: string;
+}
+
+export interface IChooseBestPhysicalAttackOptions {
+  canReachForCharge?: boolean;
+  hexesMoved?: number;
+  isJumping?: boolean;
+  hasMeleeWeapon?: PhysicalAttackType;
+  weaponsFiredFromLeftArm?: readonly string[];
+  weaponsFiredFromRightArm?: readonly string[];
+  attackerProne?: boolean;
+  heat?: number;
+  hasTSM?: boolean;
+}
+
+export interface IPhysicalAttackCandidate {
+  type: PhysicalAttackType;
+  expectedDamage: number;
+}
+
+export interface IPhysicalAttackUnitStatePair {
+  attacker: IUnitGameState;
+  target: IUnitGameState;
+}
